Fixed [SEE-872]: SubEthaEditHelperTool fails with FileVault
Starting with Mac OS X 10.4, FindFolder returns a folder in your home directory
when you look for a temp folder in the user domain. But with FileVault turned
on, your home directory is mounted from a disk image with nosuid.
We now use the kOnSystemDisk paramter to get a folder in the system disk where
setuid should be allowed.