[FIX #62 #111] Improve bash mode and fix indentation  (#140)
* First step towards better bash mode. indenting also needs see behavioral update imho
* [FIX #62] Fix indentation for else and elif in bash

fixed by making the if state in bash end at the newline before the else or elif - and also by not inspecting the newline character for determining the min indent for a line (as we don't want this state to reduce the indent. We probably should generalize this even more to not report the state end's indentation if there is content before. e.g.

if (foo) {
  something
  smoething else }

should not outdent the last line.
* [XCODE] Adding xcdebugger to the ignore list and removing the accidentially shared breakpoints